Extract `CatsInstances` to its own trait
