+++ src/main/scala/memeid/UUID.scala
@@ -1,12 +1,10 @@
 package memeid
 
-import java.lang.Long.compareUnsigned
 import java.util.{UUID => JUUID}
 
 import scala.reflect.ClassTag
 
 import cats.Show
-import cats.instances.uuid._
 import cats.kernel._
 
 /**
@@ -78,7 +76,7 @@ sealed trait UUID {
 
 }
 
-object UUID extends Constructors {
+object UUID extends Constructors with CatsInstances {
 
   /**
    * The nil UUID is special form of UUID that is specified to have all 128 bits set to zero.
@@ -134,19 +132,5 @@ object UUID extends Constructors {
       val version: Int,
       override private[memeid] val juuid: JUUID
   ) extends UUID
-
-  implicit val UUIDHashOrderShowInstance: Order[UUID] with Hash[UUID] with Show[UUID] =
-    new Order[UUID] with Hash[UUID] with Show[UUID] {
-
-      override def hash(x: UUID): Int = Hash[JUUID].hash(x.juuid)
-
-      override def compare(x: UUID, y: UUID): Int =
-        compareUnsigned(x.msb, y.msb) match {
-          case 0 => compareUnsigned(x.lsb, y.lsb)
-          case x => x
-        }
-
-      override def show(t: UUID): String = Show[JUUID].show(t.juuid)
-    }
 
 }
